LayoutTests/platform/chromium/accessibility/scroll-to-make-visible-iframe-expected.txt

Tests that scrolling to make an element visible successfully scrolls an iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS frameWindow.pageYOffset is 0
PASS frameWindow.pageYOffset is >= 4910
PASS 5038 is >= frameWindow.pageYOffset
PASS frameWindow.pageYOffset is >= -118
PASS 10 is >= frameWindow.pageYOffset
PASS successfullyParsed is true

TEST COMPLETE
